f2c: Changed implementation of c_abs(complex *) from double to float

+++ SRC/cgeqpf.c
@@ -29,7 +29,7 @@ static integer c__1 = 1;
     /* Builtin functions */
     float sqrtf(real);
     void r_cnjg(complex *, complex *);
-    double c_abs(complex *);
+    float c_abs(complex *);
 
     /* Local variables */
     integer i__, j, ma, mn;
